refactor(web): flatten home feature structure
