+++ web/features/home/banner/__tests__/indicator-button.spec.tsx
