+++ web/features/home/home-content/home-content.tsx
@@ -37,7 +37,7 @@ import { trackCreateApp } from '@/utils/create-app-tracking'
 import { hasPermission } from '@/utils/permission'
 import { HomeBanner } from '../banner/home-banner'
 import { HomeShell } from '../home-shell'
-import { TemplateCard } from '../template-card/template-card'
+import { TemplateCard } from '../template-card'
 import { HomeRecommendations } from './recommendations'
 import s from './style.module.css'
 import { HomeTemplatesHeader } from './templates-header'
